+++ NEWS.md
@@ -1,8 +1,16 @@
+# vazul 0.4.0
+## 2025-09-29
+
+* Add `together` parameter to `scramble_variables()` function.
+* Update documentation
+* Add new tests
+
 # vazul 0.3.0
 ## 2025-09-25
 
 * Add `mask_labels()`, `mask_variables()`, `mask_labels_rowwise()`, `mask_variables_rowwise()`, `mask_names()` functions to mask factor levels and variable names.
 * Add unit tests for all new functions
+* Add lifcycle badge to README
 
 # vazul 0.2.1
 ## 2025-09-19